chore: Support for trying to connect to one of multiple nodes, minor improvements and fixes
